Move  getAllMandatoryTrainings end point and model to training service and model
