+++ src/app/features/add-mandatory-training/add-mandatory-training.component.ts
@@ -2,14 +2,9 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormArray, FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ErrorDefinition, ErrorDetails } from '@core/model/errorSummary.model';
-import {
-  allMandatoryTrainingCategories,
-  Establishment,
-  mandatoryTrainingCategories,
-  mandatoryTrainingJobOption,
-} from '@core/model/establishment.model';
+import { Establishment, mandatoryTrainingCategories, mandatoryTrainingJobOption } from '@core/model/establishment.model';
 import { Job } from '@core/model/job.model';
-import { TrainingCategory } from '@core/model/training.model';
+import { allMandatoryTrainingCategories, TrainingCategory } from '@core/model/training.model';
 import { URLStructure } from '@core/model/url.model';
 import { AlertService } from '@core/services/alert.service';
 import { BackService } from '@core/services/back.service';
@@ -82,7 +77,7 @@ export class AddMandatoryTrainingComponent implements OnInit, OnDestroy {
     this.setupFormErrorsMap();
     this.setupServerErrorsMap();
     this.subscriptions.add(
-      this.establishmentService.getAllMandatoryTrainings(this.establishment.uid).subscribe(
+      this.trainingService.getAllMandatoryTrainings(this.establishment.uid).subscribe(
         (trainings) => {
           this.existingMandatoryTrainings = trainings;
           this.prefill(trainings);
